requirements.txt
transformers
torch
numpy
scikit-learn
faiss-cpu  # Or faiss-gpu if you have CUDA